renamed files specific to søknadsoppsummering
