+++ src/pages/søknad/steg/oppsummering/Søknadoppsummering/søknadsoppsummering-nb.ts
